favorite fixes
