+++ pages/api/resolvers.ts
@@ -127,14 +127,6 @@ const resolvers = {
         return { token: token };
       });
     },
-    logout: async (_, __, { res }) => {
-      try {
-        res.deleteHeader("Set-Cookie");
-      } catch {
-        return false;
-      }
-      return true;
-    },
     signupUser: (_, { username, email, password }, { dataBase, res }) => {
       return dataBase.Users.create({
         username: username,
